pyproject.toml
[project]
name = "mcp-oda"
version = "0.1.0"
description = "Integration with Oda for searching products and managing your shopping cart."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "mcp[cli]==1.10.1",
  "playwright==1.53.0",
]

[build-system]
requires = [
  "setuptools",
  "setuptools-scm",
]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["mcp_oda"]

[project.scripts]
mcp-oda = "mcp_oda.__main__:main"

[dependency-groups]
dev = [
  "icecream>=2.1.4",
  "pyright>=1.1.400",
  "pytest-playwright-asyncio>=0.7.0",
  "pytest-xdist>=3.8.0",
  "ruff>=0.11.9",
]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "E501",
  "D",
  "S603",
  "S607",
  "INP001",
  "PGH003",
  "PLR2004",
  "S101",
  "ERA001",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_test_loop_scope = "session"
asyncio_default_fixture_loop_scope = "session"
addopts = "--browser chromium --durations=0 -n auto"
# addopts = "--headed --browser chromium"
testpaths = ["tests"]
